MCP/requirements.txt
# MCP 데스크탑 앱 의존성

# 오디오 처리
pyaudio>=0.2.13          # 오디오 녹음/재생
pynput>=1.7.6            # 핫키 처리

# 네트워크 통신
websockets>=12.0         # WebSocket 클라이언트
aiohttp>=3.9.0           # 비동기 HTTP 클라이언트

# 브라우저 자동화
playwright>=1.40.0       # 브라우저 자동화 (MCP 도구용)
psutil>=5.9.0            # 프로세스 관리

# UI
pystray>=0.19.0          # 시스템 트레이 아이콘
Pillow>=10.0.0           # 이미지 처리 (아이콘용)

# 설정 관리
python-dotenv>=1.0.0     # 환경 변수 로드

# 테스트
pytest>=7.4.0            # 테스트 프레임워크
pytest-asyncio>=0.21.0   # 비동기 테스트
pytest-cov>=4.1.0        # 테스트 커버리지

# 개발 도구
black>=23.0.0            # 코드 포매터
flake8>=6.0.0            # 코드 린터
mypy>=1.5.0              # 타입 체커

# 배포
pyinstaller>=6.0.0       # .exe 패키징
